# forum/src/main/resources/documentacao Swagger.yaml
openapi: 3.0.1
info:
  title: API de cadastro de estados
  description: API para controlar Estados e suas características populacionais.
  version: 0.0.1
  termsOfService: https://example.com.br
  contact:
    name: Suporte para duvidas sobre a aplicação
    email: bruno.seixa.pb@compasso.com.br
  license:
    name: Licença:GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: https://localhost:8080
    description: API Estados
paths:
  /api/states/{id}:
    parameters:
      - name: id
        in: path
        schema:
          type: integer
        required: true
    get:
      summary: Recupera um Estado pelo ID
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  regiao:
                    type: string
                  populacao:
                    type: integer
                  capital:
                    type: string
                  area:
                    type: string
                  Nome:
                    type: string
        '404':
          description: Entidade não encontrada
          content:
            application/json:
              example: Not found
    delete:
      summary: Apaga estados pelo id
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '404':
          description: id not found
  /api/states:
    get:
      summary: Retorna todos os valores da entidade Estado
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  regiao:
                    type: string
                  populacao:
                    type: integer
                  capital:
                    type: string
                  area:
                    type: string
                  Nome:
                    type: string
    post:
      summary: Cria novo Estado
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                regiao:
                  type: string
                populacao:
                  type: integer
                capital:
                  type: string
                area:
                  type: string
                Nome:
                  type: string
      responses:
        '201':
          description: Sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  regiao:
                    type: string
                  populacao:
                    type: integer
                  capital:
                    type: string
                  area:
                    type: string
                  Nome:
                    type: string
  /api/states/populacaoOrdenada:
    get:
      summary: >-
        Retorna todos os valores da entidade Estado com o atributo populacao
        ordenado
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  regiao:
                    type: string
                  populacao:
                    type: integer
                  capital:
                    type: string
                  area:
                    type: string
                  Nome:
                    type: string
  /api/states/areaOrdenada:
    get:
      summary: Retorna todos os valores da entidade Estado com o atributo area ordenado
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  regiao:
                    type: string
                  populacao:
                    type: integer
                  capital:
                    type: string
                  area:
                    type: string
                  Nome:
                    type: string